Add generate / scafolding commands, templates based on snippets syntax
Basically placeholders are parsed to determine which values are going to
be prompted. Once done (prompts done), the according directory is
created (routers, views or models, todo: this should be configurable),
and a new buffer is created with the content of the existsting file (if
any), or creates a new one (a simple :edit ./pathto/file.js).

The template content is then appended to the current buffer.

:BRouter, :BModel, :BCollection, :BView and :BGenerate (the later is the
actual generation command, others are simple wrappers to :BGenerate)